[Feat/#290] :wrench: Refactor: 다른 곳에서 쓰이지 않는 record를 관련 class 안으로 이동
