--- chatty-be/src/main/java/com/chatty/chatty/game/controller/dto/model/AnswerDTO.java
@@ -1,10 +1,0 @@
-package com.chatty.chatty.game.controller.dto.model;
-
-import lombok.Builder;
-
-@Builder
-public record AnswerDTO(
-        Long user_id,
-        String user
-) {
-}
