--- chatty-be/src/main/java/com/chatty/chatty/game/domain/PlayerAnswerData.java
@@ -1,12 +1,0 @@
-package com.chatty.chatty.game.domain;
-
-import java.time.LocalDateTime;
-import lombok.Builder;
-
-@Builder
-public record PlayerAnswerData(
-        String playerAnswer,
-
-        LocalDateTime submittedTime
-) {
-}
