+++ chatty-be/src/main/java/com/chatty/chatty/game/domain/ScoreData.java
@@ -1,0 +1,9 @@
+package com.chatty.chatty.game.domain;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class ScoreData {
+}
