--- chatty-be/src/main/java/com/chatty/chatty/game/service/ScoreService.java
@@ -1,4 +1,0 @@
-package com.chatty.chatty.game.service;
-
-public class ScoreService {
-}
